# Week3/Day4/openapi.yaml
openapi: 3.0.0
info:
  title: Tasks API
  description: A simple versioned Flask API for managing tasks
  version: "1.0.0"
servers:
  - url: http://127.0.0.1:5000/api/v1
paths:
  /tasks:
    get:
      summary: Get all tasks
      responses:
        '200':
          description: A list of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    title:
                      type: string
                    completed:
                      type: boolean
    post:
      summary: Add a new task
      requestBody:
        description: Task object to add
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                completed:
                  type: boolean
                  default: false
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  title:
                    type: string
                  completed:
                    type: boolean
  /tasks/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: ID of the task
        schema:
          type: integer
    put:
      summary: Update a task by ID
      requestBody:
        description: Updated task object
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                completed:
                  type: boolean
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  title:
                    type: string
                  completed:
                    type: boolean
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
    delete:
      summary: Delete a task by ID
      responses:
        '200':
          description: Task deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: Task not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string